Modify makefiles and setup.conf for riscv-toolchain
This patch adds modified config files to support building
cross-compilation toolchain for riscv. It will create a
tools/riscv dir to host the toolchains similar to the current
configuration for superH.

It is also worth noting that riscv is only supported by higher
versions of gcc and binutils. For binutils, it seems that versions
supporting riscv have already dropped superH support.